Restructure all apps into one folder
